+++ src/test/java/step5/controller/RacingCarControllerTest.java
@@ -1,29 +1,28 @@
-package step5;
+package step5.controller;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import step5.domain.Car;
-import step5.domain.RacingCarGame;
 
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class RacingCarGameTest {
+public class RacingCarControllerTest {
 
-    RacingCarGame racingCarGame;
+    RacingCarController racingCarController;
 
     @BeforeEach
     void setUp() {
-        racingCarGame = new RacingCarGame();
+        racingCarController = new RacingCarController();
     }
 
     @Test
     @DisplayName("입력한 자동차 갯수 일치 확인")
     public void 입력한_자동차_갯수_일치_확인() {
         String[] carNames = "zu,cu,du".split(",");
-        List<Car> cars = racingCarGame.start(carNames);
+        List<Car> cars = racingCarController.start(carNames);
         assertEquals(carNames.length, cars.size());
     }
 
@@ -31,7 +30,7 @@ public class RacingCarGameTest {
     @DisplayName("입력한 자동차 값이 정상 등록되는지")
     public void 등록_자동차_확인() {
         String[] carNames = "zu,cu,du".split(",");
-        List<Car> cars = racingCarGame.start(carNames);
+        List<Car> cars = racingCarController.start(carNames);
         assertEquals(cars.get(0).getCarName(), "zu");
         assertEquals(cars.get(1).getCarName(), "cu");
         assertEquals(cars.get(2).getCarName(), "du");
