+++ src/test/java/step5/domain/RacingWinnerTest.java
@@ -1,9 +1,7 @@
-package step5;
+package step5.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step5.domain.Car;
-import step5.domain.RacingWinner;
 
 import java.util.ArrayList;
 import java.util.List;
